utils.py, dataset_modelling.py, updated demo_notebook
Features added:

1. Printing out the denormalized data in prediction - DONE
2. Store the ids of NaNs before filling them up with placeholders - Not Needed, wrote a better test_model function
3. Do 70/30 Split before training, testing - DONE
4. Add provision in the model for changing theta params - DONE
5. Add provision for removing dropout - DONE
6. Plotting Loss function - DONE

              
Experiments to try:
1. Plot different values for different architectures.
2. Try feeding in with just the imputed values instead of the whole predicted data.
3. Try without dropout.

+++ Modelling.py
@@ -28,7 +28,7 @@ class DenoisingAutoEncoder(nn.Module):
             self.linear_layer_list.append(nn.Linear(*in_out_pair))
 
     def forward(self, x):
-
+        x = x.float()
         h = self.drop_layer(x)
         for layer in self.linear_layer_list[:-1]:
             h = torch.tanh(layer(h))
